requirements.txt
# Step 1: Install PyTorch first
# pip install torch==2.1.2 torchvision==0.16.2 torchaudio==2.1.2 --index-url https://download.pytorch.org/whl/cu118

# Step 2: Install PyTorch Geometric
torch-geometric==2.4.0

# Data processing and scientific computing
numpy>=1.19.0
pandas>=1.3.0

# Visualization
matplotlib>=3.3.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.62.0

# Optional: for better performance
scipy>=1.7.0

# Fix dependency conflicts
filelock>=3.0.0
oauthlib>=3.0.0
opencv-python>=4.5.5.64